Separate pure and side-effecting code
