+++ lib/helpers.js
@@ -132,9 +132,52 @@ function renameCache(source, destination, options) {
   });
 }
 
+function cache(url, options) {
+  return openCache(options).then(function(cache) {
+    return cache.add(url);
+  });
+}
+
+function uncache(url, options) {
+  return openCache(options).then(function(cache) {
+    return cache.delete(url);
+  });
+}
+
+function precache(items) {
+  if (!(items instanceof Promise)) {
+    validatePrecacheInput(items);
+  }
+
+  globalOptions.preCacheItems = globalOptions.preCacheItems.concat(items);
+}
+
+function validatePrecacheInput(items) {
+  var isValid = Array.isArray(items);
+  if (isValid) {
+    items.forEach(function(item) {
+      if (!(typeof item === 'string' || (item instanceof Request))) {
+        isValid = false;
+      }
+    });
+  }
+
+  if (!isValid) {
+    throw new TypeError('The precache method expects either an array of ' +
+    'strings and/or Requests or a Promise that resolves to an array of ' +
+    'strings and/or Requests.');
+  }
+
+  return items;
+}
+
 module.exports = {
   debug: debug,
   fetchAndCache: fetchAndCache,
   openCache: openCache,
-  renameCache: renameCache
+  renameCache: renameCache,
+  cache: cache,
+  uncache: uncache,
+  precache: precache,
+  validatePrecacheInput: validatePrecacheInput
 };
