+++ lib/sw-toolbox.js
@@ -15,103 +15,22 @@
 */
 'use strict';
 
-require('serviceworker-cache-polyfill');
+// This is the entrypoint for the sw-toolbox bundle. All code with
+// side effects (e.g. adding event listeners) should be in this file.
+
 var options = require('./options');
 var router = require('./router');
 var helpers = require('./helpers');
 var strategies = require('./strategies');
+var listeners = require('./listeners');
 
 helpers.debug('Service Worker Toolbox is loading');
 
-// Install
-var flatten = function(items) {
-  return items.reduce(function(a, b) {
-    return a.concat(b);
-  }, []);
-};
+// Set up listeners.
 
-var validatePrecacheInput = function(items) {
-  var isValid = Array.isArray(items);
-  if (isValid) {
-    items.forEach(function(item) {
-      if (!(typeof item === 'string' || (item instanceof Request))) {
-        isValid = false;
-      }
-    });
-  }
-
-  if (!isValid) {
-    throw new TypeError('The precache method expects either an array of ' +
-    'strings and/or Requests or a Promise that resolves to an array of ' +
-    'strings and/or Requests.');
-  }
-
-  return items;
-};
-
-self.addEventListener('install', function(event) {
-  var inactiveCache = options.cache.name + '$$$inactive$$$';
-  helpers.debug('install event fired');
-  helpers.debug('creating cache [' + inactiveCache + ']');
-  event.waitUntil(
-    helpers.openCache({cache: {name: inactiveCache}})
-    .then(function(cache) {
-      return Promise.all(options.preCacheItems)
-      .then(flatten)
-      .then(validatePrecacheInput)
-      .then(function(preCacheItems) {
-        helpers.debug('preCache list: ' +
-            (preCacheItems.join(', ') || '(none)'));
-        return cache.addAll(preCacheItems);
-      });
-    })
-  );
-});
-
-// Activate
-
-self.addEventListener('activate', function(event) {
-  helpers.debug('activate event fired');
-  var inactiveCache = options.cache.name + '$$$inactive$$$';
-  event.waitUntil(helpers.renameCache(inactiveCache, options.cache.name));
-});
-
-// Fetch
-
-self.addEventListener('fetch', function(event) {
-  var handler = router.match(event.request);
-
-  if (handler) {
-    event.respondWith(handler(event.request));
-  } else if (router.default &&
-    event.request.method === 'GET' &&
-    // Ensure that chrome-extension:// requests don't trigger the default route.
-    event.request.url.indexOf('http') === 0) {
-    event.respondWith(router.default(event.request));
-  }
-});
-
-// Caching
-
-function cache(url, options) {
-  return helpers.openCache(options).then(function(cache) {
-    return cache.add(url);
-  });
-}
-
-function uncache(url, options) {
-  return helpers.openCache(options).then(function(cache) {
-    return cache.delete(url);
-  });
-}
-
-function precache(items) {
-  if (!(items instanceof Promise)) {
-    validatePrecacheInput(items);
-  }
-
-  options.preCacheItems = options.preCacheItems.concat(items);
-}
+self.addEventListener('install', listeners.installListener);
+self.addEventListener('activate', listeners.activateListener);
+self.addEventListener('fetch', listeners.fetchListener);
 
 module.exports = {
   networkOnly: strategies.networkOnly,
@@ -121,7 +40,7 @@ module.exports = {
   fastest: strategies.fastest,
   router: router,
   options: options,
-  cache: cache,
-  uncache: uncache,
-  precache: precache
+  cache: helpers.cache,
+  uncache: helpers.uncache,
+  precache: helpers.precache
 };
